+++ packages/engine/src/scene/components/ObjectLayerComponent.test.tsx
@@ -26,17 +26,13 @@ Ethereal Engine. All Rights Reserved.
 import assert from 'assert'
 import { BoxGeometry, Mesh, MeshBasicMaterial } from 'three'
 
-import { destroyEngine } from '../../../src/ecs/classes/Engine'
-import { getComponent, hasComponent, setComponent } from '../../../src/ecs/functions/ComponentFunctions'
-import { createEntity } from '../../../src/ecs/functions/EntityFunctions'
-import { createEngine } from '../../../src/initializeEngine'
-import { addObjectToGroup } from '../../../src/scene/components/GroupComponent'
-import {
-  Layer,
-  ObjectLayerComponents,
-  ObjectLayerMaskComponent
-} from '../../../src/scene/components/ObjectLayerComponent'
-import { loadEmptyScene } from '../../util/loadEmptyScene'
+import { loadEmptyScene } from '../../../tests/util/loadEmptyScene'
+import { destroyEngine } from '../../ecs/classes/Engine'
+import { getComponent, hasComponent, setComponent } from '../../ecs/functions/ComponentFunctions'
+import { createEntity } from '../../ecs/functions/EntityFunctions'
+import { createEngine } from '../../initializeEngine'
+import { addObjectToGroup } from './GroupComponent'
+import { Layer, ObjectLayerComponents, ObjectLayerMaskComponent } from './ObjectLayerComponent'
 
 describe('ObjectLayerComponent', () => {
   beforeEach(async () => {
@@ -61,7 +57,7 @@ describe('ObjectLayerComponent', () => {
     mesh.layers.enable(objectLayer)
 
     assert(hasComponent(entity, ObjectLayerMaskComponent))
-    assert(hasComponent(entity, ObjectLayerComponents[objectLayer + 1]))
+    assert(hasComponent(entity, ObjectLayerComponents[objectLayer]))
     assert(!hasComponent(entity, ObjectLayerComponents[nonEnabledObjectLayer]))
     assert(mesh.layers.isEnabled(objectLayer))
     assert(!mesh.layers.isEnabled(nonEnabledObjectLayer))
@@ -89,7 +85,7 @@ describe('ObjectLayerComponent', () => {
     for (const mesh of meshes) {
       for (const layer of objectLayers) {
         assert(mesh.layers.isEnabled(layer))
-        assert(hasComponent(entity, ObjectLayerComponents[layer + 1]))
+        assert(hasComponent(entity, ObjectLayerComponents[layer]))
       }
     }
   })
@@ -109,7 +105,7 @@ describe('ObjectLayerComponent', () => {
 
     for (const layer of objectLayers) {
       assert(mesh.layers.isEnabled(layer))
-      assert(hasComponent(entity, ObjectLayerComponents[layer + 1]))
+      assert(hasComponent(entity, ObjectLayerComponents[layer]))
     }
 
     assert(!mesh.layers.isEnabled(nonEnabledObjectLayer))
@@ -120,10 +116,10 @@ describe('ObjectLayerComponent', () => {
     for (const layer of objectLayers) {
       if (disableLayers.includes(layer)) {
         assert(!mesh.layers.isEnabled(layer))
-        assert(!hasComponent(entity, ObjectLayerComponents[layer + 1]))
+        assert(!hasComponent(entity, ObjectLayerComponents[layer]))
       } else {
         assert(mesh.layers.isEnabled(layer))
-        assert(hasComponent(entity, ObjectLayerComponents[layer + 1]))
+        assert(hasComponent(entity, ObjectLayerComponents[layer]))
       }
     }
   })
@@ -138,7 +134,7 @@ describe('ObjectLayerComponent', () => {
     const layerMaskComponent = getComponent(entity, ObjectLayerMaskComponent)
     assert(layerMaskComponent === 1)
     assert(ObjectLayerMaskComponent.mask[entity] === 1)
-    assert(hasComponent(entity, ObjectLayerComponents[1]))
+    assert(hasComponent(entity, ObjectLayerComponents[0]))
 
     const enabledLayers = [2, 3, 4]
 
@@ -158,14 +154,14 @@ describe('ObjectLayerComponent', () => {
     assert(layer.mask === 0)
     for (let i = 0; i < maxLayers; i++) {
       assert(!layer.isEnabled(i))
-      assert(!hasComponent(entity, ObjectLayerComponents[i + 1]))
+      assert(!hasComponent(entity, ObjectLayerComponents[i]))
     }
 
     layer.enableAll()
     assert(layer.mask.valueOf() === (0xffffffff | 0))
     for (let i = 1; i < maxLayers; i++) {
       assert(layer.isEnabled(i))
-      assert(hasComponent(entity, ObjectLayerComponents[i + 1]))
+      assert(hasComponent(entity, ObjectLayerComponents[i]))
     }
 
     layer.toggle(4)
